@@ -1,19 +1,18 @@
 import { NgStyle } from '@angular/common';
-import { Component, Input, OnInit, SimpleChanges } from '@angular/core';
+import { Component, OnInit, input } from '@angular/core';
 
 @Component({
-  selector: 'app-hashtags',
-  standalone: true,
-  imports: [NgStyle],
-  templateUrl: './hashtags.component.html',
-  styleUrl: './hashtags.component.scss'
+    selector: 'app-hashtags',
+    imports: [NgStyle],
+    templateUrl: './hashtags.component.html',
+    styleUrl: './hashtags.component.scss'
 })
 export class HashtagsComponent implements OnInit {
-  @Input() hashtags: string[] = [];
+  readonly hashtags = input<string[]>([]);
   hashtagGradients: string[] = [];
 
   ngOnInit(): void {
-    this.hashtagGradients = this.hashtags.map(() => this.getRandomGradient());
+    this.hashtagGradients = this.hashtags().map(() => this.getRandomGradient());
   }
 
   /**
