@@ -1,17 +1,16 @@
-import { Component, EventEmitter, Input, Output } from '@angular/core';
+import { Component, EventEmitter, Output, input } from '@angular/core';
 
 @Component({
-  selector: 'app-popup',
-  standalone: true,
-  imports: [],
-  templateUrl: './popup.component.html',
-  styleUrl: './popup.component.scss'
+    selector: 'app-popup',
+    imports: [],
+    templateUrl: './popup.component.html',
+    styleUrl: './popup.component.scss'
 })
 export class PopupComponent {
-  @Input() isOpen: boolean = false;
+  readonly isOpen = input<boolean>(false);
   @Output() toggle = new EventEmitter<boolean>();
 
   closePopup() {
-    this.toggle.emit(!this.isOpen);
+    this.toggle.emit(!this.isOpen());
   }
 }
