@@ -1,11 +1,10 @@
 import { Component } from '@angular/core';
 
 @Component({
-  selector: 'app-recognitions-contributions',
-  standalone: true,
-  imports: [],
-  templateUrl: './recognitions-contributions.component.html',
-  styleUrl: './recognitions-contributions.component.scss'
+    selector: 'app-recognitions-contributions',
+    imports: [],
+    templateUrl: './recognitions-contributions.component.html',
+    styleUrl: './recognitions-contributions.component.scss'
 })
 export class RecognitionsContributionsComponent {
 
