@@ -1,11 +1,10 @@
 import { Component } from '@angular/core';
 
 @Component({
-  selector: 'app-certifications',
-  standalone: true,
-  imports: [],
-  templateUrl: './certifications.component.html',
-  styleUrl: './certifications.component.css'
+    selector: 'app-certifications',
+    imports: [],
+    templateUrl: './certifications.component.html',
+    styleUrl: './certifications.component.css'
 })
 export class CertificationsComponent {
 
