@@ -1,11 +1,10 @@
 import { Component } from '@angular/core';
 
 @Component({
-  selector: 'app-honors-awards',
-  standalone: true,
-  imports: [],
-  templateUrl: './honors-awards.component.html',
-  styleUrl: './honors-awards.component.css'
+    selector: 'app-honors-awards',
+    imports: [],
+    templateUrl: './honors-awards.component.html',
+    styleUrl: './honors-awards.component.css'
 })
 export class HonorsAwardsComponent {
 
