@@ -1,11 +1,10 @@
 import { Component } from '@angular/core';
 
 @Component({
-  selector: 'app-publications',
-  standalone: true,
-  imports: [],
-  templateUrl: './publications.component.html',
-  styleUrl: './publications.component.css'
+    selector: 'app-publications',
+    imports: [],
+    templateUrl: './publications.component.html',
+    styleUrl: './publications.component.css'
 })
 export class PublicationsComponent {
 
